core/requirements.txt
asgiref==3.7.2
certifi==2023.5.7
cffi==1.15.1
channels==4.0.0
charset-normalizer==3.2.0
click==8.1.3
colorama==0.4.6
crispy-bootstrap5==0.7
cryptography==41.0.4
cssbeautifier==1.14.8
Django==4.2.6
django-crispy-forms==2.0
django-rest-knox==4.2.0
django-restframework==0.0.1
djangorestframework==3.14.0
EditorConfig==0.12.3
html-tag-names==0.1.2
html-void-elements==0.1.0
idna==3.4
jsbeautifier==1.14.8
json5==0.9.14
pathspec==0.11.1
Pillow==9.5.0
psycopg2==2.9.6
pycparser==2.21
python-decouple==3.8
pytz==2023.3.post1
PyYAML==6.0
regex==2023.6.3
requests==2.31.0
simplejson==3.19.1
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
tqdm==4.65.0
typing_extensions==4.6.3
tzdata==2023.3
urllib3==2.0.3
